Update unit tests. Now they should be separated and run in insolated classes.
