+++ test/py2sql/test_save_object.py
@@ -1,27 +1,13 @@
 import unittest
+
+import unittest
 import sys
 import os
-sys.path.append('../..')
+sys.path.insert(0, os.getcwd())
 from py2sql import py2sql
 
-class Sample:
-    foo: int
-    bar: str
-    def __init__(self):
-        self.foo = 1
-        self.bar = "bar"
 
-class SubSample(Sample):
-    zoo: int
-    def __init__(self):
-        self.zoo = 2
-
-class Bar:
-    done: bool
-    def __init__(self):
-        self.done = True
-
-class TestSavingHierarchy(unittest.TestCase):
+class F(unittest.TestCase):
 
     db_config = py2sql.DBConnectionInfo("test", "localhost", "adminadminadmin", "postgres")
 
@@ -30,21 +16,14 @@ class TestSavingHierarchy(unittest.TestCase):
         py2sql.Py2SQL.drop_table("test")
         py2sql.Py2SQL.db_disconnect()
 
-    def test_save_class_normal_usecase(self):
-        py2sql.Py2SQL.db_connect(self.db_config)
-        py2sql.Py2SQL.save_class(Bar)
-        self.assertNotEqual((), py2sql.Py2SQL.db_table_structure("bar"))
-        py2sql.Py2SQL.save_class(Bar)
-        self.assertNotEqual((), py2sql.Py2SQL.db_table_structure("bar"))
-        py2sql.Py2SQL.db_disconnect()
-
-    def test_save_object_normal_usecase(self):
+    def save_object_with_no_class_raises_exception():
         py2sql.Py2SQL.db_connect(self.db_config)
         b = Bar()
         py2sql.Py2SQL.save_object(b)
         py2sql.Py2SQL.db_disconnect()
 
-    def test_integration(self):
+    def save_class_and_object():
+        py2sql.Py2SQL.db_connect(self.db_config)
 
         class S:
             foo: int
@@ -60,6 +39,4 @@ class TestSavingHierarchy(unittest.TestCase):
         py2sql.Py2SQL.save_object(s)
         py2sql.Py2SQL.db_disconnect()
 
-
-if __name__ == "__main__":
-    unittest.main()
+        py2sql.Py2SQL.db_disconnect()
